+++ app/components/course-admin/submissions-page/submission-details/community-solution-container.hbs
@@ -1,4 +1,3 @@
-{{! @glint-nocheck: not typesafe yet }}
 {{#if @submission.communityCourseStageSolution}}
   <CoursePage::CourseStageStep::CommunitySolutionCard @solution={{@submission.communityCourseStageSolution}} @isCollapsedByDefault={{true}} />
 {{else}}
