--- app/components/course-admin/submissions-page/submission-details/community-solution-container.js
@@ -1,3 +1,0 @@
-import Component from '@glimmer/component';
-
-export default class CommunitySolutionContainerComponent extends Component {}
