+++ app/components/course-admin/submissions-page/submission-details/diff-container.hbs
@@ -1,20 +1,24 @@
-{{! @glint-nocheck: not typesafe yet }}
 <div ...attributes>
   {{#if @submission.hasChangedFiles}}
     {{#each @submission.changedFiles key="filename" as |changedFile index|}}
       <div
-        class="{{if (not-eq index (sub @solution.changedFiles.length 1)) 'mb-4'}} bg-white shadow-sm rounded border border-gray-300 overflow-y-hidden"
+        class="{{if (not-eq index (sub @submission.changedFiles.length 1)) 'mb-4'}}
+          bg-white shadow-sm rounded border border-gray-300 overflow-y-hidden"
       >
         <div class="bg-gray-100 rounded-t py-1 px-4 border-b border-gray-200 shadow-sm">
           <span class="font-mono text-xs text-gray-600 bold">{{changedFile.filename}}</span>
         </div>
 
         <div class="text-sm leading-6">
-          <SyntaxHighlightedDiff
-            @code={{changedFile.diff}}
-            @language={{@submission.repository.language.slug}}
-            @shouldCollapseUnchangedLines={{true}}
-          />
+          {{! #if ensures that language is typed and not null }}
+          {{#if @submission.repository.language}}
+            <SyntaxHighlightedDiff
+              @comments={{array}}
+              @code={{changedFile.diff}}
+              @language={{@submission.repository.language.slug}}
+              @shouldCollapseUnchangedLines={{true}}
+            />
+          {{/if}}
         </div>
       </div>
     {{/each}}
