--- app/components/course-admin/submissions-page/submission-details/diff-container.js
@@ -1,4 +1,0 @@
-import Component from '@glimmer/component';
-
-// eslint-disable-next-line ember/no-empty-glimmer-component-classes
-export default class DiffContainerComponent extends Component {}
